supreme-commander-agent/pyproject.toml
[project]
name = "supreme-commander-agent"
version = "1.0.0"
description = "Supreme Commander - Boss of all boss agents for multi-agent orchestration"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.60.0",
    "google-adk-python>=0.1.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "firebase-admin>=6.5.0",
    "vertexai>=1.60.0",
    "google-cloud-discoveryengine>=0.11.0",
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.16.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-gcp-trace>=1.6.0",
    "pydantic>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
    "codespell>=2.3.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
